refactor(cayenne): simplify filtered maintained-aggregate apply, share the bench fixture, document the serve-match boundary
/simplify cleanup of the predicate-aware maintained-aggregate change; no behavior change.

- maintained_aggregate.rs: collapse `apply_insert_batch`'s indexed/unindexed branches
  into one shared tail — compute the PK once, then a single match-gate + group insert —
  instead of duplicating the post-filter path (`continue`, group-key, insert) in both arms.
- Document the serve-match boundary on `matches_query` and the runtime filter parser:
  the match is structural `PhysicalExpr` equality over the scan-output schema, so it
  silently does NOT fire across a projection / cast / type-coercion between the scan and
  the filter (e.g. `SchemaCastScanExec`, or `Utf8View`-over-`Utf8`) and falls back to a
  re-scan. Name-based predicate normalization is the follow-up.
- benches: extract the duplicated Cayenne-side fixture (schema, spec, row generator,
  registry load, serve decode) into `maintained_filtered_helpers/common.rs`, shared by the
  vs_duckdb and vs_chdb benches via the established `#[path] mod` pattern, so the
  registry/spec contract can't drift between the two split binaries.

Validated: `cargo test -p cayenne maintained_aggregate` (20 lib + 2 integration tests)
green; both maintained-filtered benches compile under duckdb-bench/chdb-bench; fmt clean.
